// src/lib/volvo/conve.gen.ts
export interface components {
    schemas: {
// ...
        VehicleDetails: {
            vin?: string;
            /** Format: int32 */
            modelYear?: number;
            gearbox?: string;
            fuelType?: string;
            externalColour?: string;
            /** Format: float */
            batteryCapacityKWH?: number;
            images?: components["schemas"]["Images"];
            descriptions?: components["schemas"]["Descriptions"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}